Remove stray warning about tasks that weren't originally `SCHEDULED`
Background tasks used to be the only tasks that existed without a flow run, but
that is no longer true.  The `TaskSchedulingTimeouts`, when it's looking for
stuck `PENDING` tasks would warn when it encountered one that hadn't previously
been `SCHEDULED`.  These tasks can just be quietly skipped here, without warning.