Require a `PREFECT_API_URL` to start a worker
When there's no `PREFECT_API_URL` set, calls to `get_client` will start the
ephemeral server transparently for the user.  This works amazingly well for
running single flows and tasks for demos and testing.  In the default
configuration, you can share a SQLite database across several Prefect processes,
like a server and a worker, and things will generally appear to work.

However, when we start to get into more advanced infrastructure like `Worker`
and `TaskWorker`, it doesn't make sense to run these ephemerally without a
Prefect server.  Event-driven workflows and background task scheduling require
a running server, and we should be explicit about that fact to avoid the
possibility for confusion (like events not being emitted, or automations not
running as expected).

Fixes #13876