Generating SDK docs with mkdocs
This script produces a tree of markdown files following our module hierarchy
to the (uncommited) `docs/sdk` directory.  After they are generated, we can
`mkdocs build` or `mkdocs serve` them.  This is the first step along the path to
building the SDK docs in CI and pushing them to Netlify or Read the Docs on
`main` merges.

Part of #13974