Controls time in the task engine cache tests to avoid flakes
This introduces a new `advance_time` fixture that allows for arbitrarily
adjusting the return value of `pendulum.now` in tests.  It's similar to its
counterpart `frozen_time`, but the time will advance on its own with each call.
The function returned by the fixture allows advancing time by a specified amount
in tests.