Adds a private status endpoint to `TaskWorker`
This endpoint, at `http://127.0.0.1:4422/status` reports the current state of
the `TaskWorker`, including what its limiting parameters look like and which
task runs of each key are currently in-flight.  This is a rough-and-ready
monitoring endpoint that we can flesh out with more info and tests as we go, but
for now, we need it for load testing and reliability troubleshooting.