Closes another race condition in task waiting
There is a very narrow gap here, but it's possible nonetheless, where an event
could sneak through in the time between checking for a previously completed
task run and when we start waiting for the event to arrive on the websocket.
This closes the gap so that we make one more check just before we start waiting,
after having established the `asyncio.Event` for completion.

Part of #14098