Disables server-side `PENDING` task rescheduling by default
During my load testing, it became clear that this feature won't work as-is on
such a short timescale.  Because tasks remain `PENDING` until all of their
dependencies are resolved, it's possible that a long period of time may elapse
before those `PENDING` tasks can actually move on to `RUNNING`, and in a system
with deep dependencies, this can put the task workers in a position where they
can't keep up and are constantly getting redelivered tasks.

Rather than defaulting this to on, we should let the user opt-in to the behavior
for now, based on their workload.