Retries retrieval of persisted results from storage
With some recent changes to transactions, we are now emitting events slightly
before we write results to their storage backend.  While we should undertake a
fuller refactoring to uncouple transaction storage from individual results, we
can also get by with a few built-in graceful retries here.  This should smooth
over the issue with the race condition, as well as help with transient network
issues on remote filesystems.

Part of #14189