Refactors event causal ordering to be reusable by multiple consumers
For the triggers system, we keep track of out-of-order messages and attempt to
replay them in causal order (where an event is processed after the event that
it declares it `follows`).  This change extracts this logic to a class and
separates the bookkeeping records so that multiple consumers can keep separate
track of events.

This is largely a port of the changes in Prefect Cloud that we made last week.

Closes #14577