Incrementing run counts during client-side task orchestration
These two counters, the `flow_run_run_count` and the `run_count`, are mostly
used during retry handling, which should already be completely handled on the
client-side and may not be necessary to track for client-side orchestration.
We do, however, want to keep the server-orchestrated and client-orchestrated
behavior as similar as possible, so they are updated here for completeness.