Removes the orchestration rule and background task to timeout pending task runs
As part of the background task scheduling project, we added functionality that
will re-schedule task runs if they have been stuck pending for a period of time.
The idea here was that if a TaskWorker crashed between receiving a run and
starting it, the task would be stuck pending forever.

This initial attempt did work, but it gave the user no control over this
behavior, and no way to tune it on a task-by-task basis.  We've had it disabled
this whole time waiting to see if a better solution offers itself.  This logic
would now interfere with client-side task run orchestration, in that we would
no longer have the protection of the Pending->Pending rule to prevent tasks from
running on two workers at a time.