Adds Prometheus metrics for API and client
This adds the `prometheus_client` library as a client-side dependency.  This
is a very lightweight (<500KB compiled) library with no transitive dependencies.
I'm choosing Prometheus here as it is the current standard for metrics from the
CNCF, and because its Python client library is very light and thus suitable for
inclusion on the client-side of Prefect.

There are two "sides" that expose a Prometheus metrics endpoint:

* The Prefect server will expose it from startup if `PREFECT_API_ENABLE_METRICS`
  is on.  This endpoint is at `/api/metrics` (to avoid any potential collision
  with future UI routes).

* The Prefect client SDK will expose these from any of the following points if
  `PREFECT_CLIENT_ENABLE_METRICS` is on:

  * Serving one or more flow via `serve(...)`
  * Serving one or more tasks via `serve(...)`
  * Entering a flow run context for a flow (e.g. entering the engine for a flow
    as a Kubernetes job)

  The client metrics will be served via `localhost` at the port defined in
  `PREFECT_CLIENT_METRICS_PORT`

Note that I haven't included any Prefect-specific metrics here, so we will start
with the stock Python-oriented metrics that come out of the box with
`prometheus_client`.