Ensures that server automations schemas are parsed with only server models
During some internal testing of Prefect, we noticed a surprising bug, where
we could not create server-side automations that included composite triggers.
This was due to Pydantic parsing the server-side `CompositeTrigger.triggers` as
_client-side_ `EventTrigger`s, which don't include some of the necessary
server-side implementation.

I was unable to reproduce the problem in a unit test (if this were happening
during unit tests, many tests would fail), but I was able to reproduce it in
a simple standalone script using the same test code.  This is hinting to
import ordering as part of the problem.

```
Traceback (most recent call last):
  File "/home/chris/src/github.com/PrefectHQ/prefect/repro.py", line 92, in <module>
    test_server_composite_triggers_validate_to_server_triggers()
  File "/home/chris/src/github.com/PrefectHQ/prefect/repro.py", line 61, in test_server_composite_triggers_validate_to_server_triggers
    assert isinstance(automation.trigger.triggers[0], EventTrigger), type(
AssertionError: <class 'prefect.events.schemas.automations.EventTrigger'>
```

Unfortunately, I was unable to create a minimal repro using just plain Pydantic
models in order to send them a bug report, but it appears to be due to some
kind of caching relating to the `TriggerTypes` union.  Renaming the union
to `ServerTriggerTypes` resolved the problem, so there must be some kind of
caching by name in Pydantic that we're triggering.