Adds acquire/release endpoints for task run concurrency limits
In support of client-side task run orchestration, we need to expose these two
endpoints so that clients may acquire and release tag-based concurrency limits
for the task runs they are orchestrating.

The design of this API is intended to be closer to V2 concurrency terminology,
hence the `increment`/`decrement` paths and the `names` parameter.