Quotes the components of a PostgreSQL connection URL when built from components
In the prior implementation, if the password had a `:` character, the URL
parsing for the connection URL would break (because it would look like
everything after the `:` was the port for the DB).  Using SQLAlchemy's built-in
URL parsing and construction ensures that each component of the URL is quoted
correctly.