Adds an `action.triggered` event in addition to the `action.executed` one (#8366)
When automations trigger, there is a single event representing that the
_automation_ triggered, then there are N additional events representing
that the actions within that automation had executed or failed. However,
because the action event was emitted after the action had finished, you can get
an unintuitive history where the action executed event follows some events
it has caused. For example, if an automation cancels a flow run, you would get
this history:

1. prefect-cloud.automation.triggered
2. prefect.flow-run.Cancelled
3. prefect-cloud.automation.action.executed

With these changes, you'll get this history:

1. prefect-cloud.automation.triggered
2. prefect-cloud.automation.action.triggered
3. prefect.flow-run.Cancelled
4. prefect-cloud.automation.action.executed (with final status
information)