Correctly handles recurring automations and future events
There is an edge case with simple recurring automations (those with an `expect`,
a `within`, no `after`, and no `for_each`).  When an event arrives with an
`occurred` time that is ahead of the `end` of the current window, we correctly
skip counting that event toward that window, and we correctly attribute the
event to the next future window.  However, we weren't testing if that new future
window would _meet its threshold_ given that event.  This would cause these
simple automations to occasionally fail to fire on schedule.

Here we make one recursive call to `evaluate` in this case where we're leaving
`evaluate` with a new bucket and a remaining `count` of events.  This ensures
that we'll immediately fire on the first event rather than relying one a steady
stream of matching events to fire the trigger.

Fixes #15343