Remove experimental worker logs setting.
Worker logs are no longer experimental, so remove references to the
setting. We still want to avoid sending worker logs to open- source API
servers until they support worker logs, though, so we use the existing
of a `backend_id` on worker logs as a proxy for whether or not the
worker is connected to Cloud. (Only workers connected to Cloud emit logs
with worker IDs attached.)