Transitions to using `traceparent` instead of separate IDs
In a prior change, we added the `trace_id` and `span_id` as separate
fields to the `StateDetails`.  However, because `trace_id` is a 128-bit
integer, we could get inconsistencies and errors serializing that to and
from JSON.  For example, browsers usually only support 53-bit integers
in JSON, which means these trace IDs become floating point values.  In
`orjson`, it strictly validates that integers are at most 64-bit and
raises an error if not.

Because of these compatibility issues, we're going to change this to
the more standard abstraction of the string-based `traceparent`, which
encodes the trace and span IDs along with additional flags.