Reduces the minimum memory requirement for Prefect Server
This brings the minimum memory requirement for Prefect Server from
~400MB to ~345MB in my testing*.

What I discovered was that when when we use FastAPI's `include_router`,
FastAPI will create _new_ routes rather than using a reference to the
previously declared router's routes.  This is because additional
dependencies may be injected that the original router wasn't aware of,
so the new routes are built using the original router as a reference.

With Prefect's (likely very commonplace) pattern of declaring
module-level routers and then assembling them into a larger FastAPI
`app`, this means that there is an entire copy of the routes left in
memory that will likely never be touched again.  Because the routes may
include dynamically generated Pydantic models representing the bodies of
requests, which themselves may be referring to other models that Prefect
defines, the complexity of the generated validators can balloon quickly.
In our case, something like ~50-55MB is held in references by those
routes.

*Note about the testing approach: I used a script that runs `prefect
server start` in a `cgroup` to constrain the memory it has access to by
setting `memory.max` and by disabling swap memory with `memory.swap=0`.
It then polls the health check for 5 consecutive successes while
confirming that the process remains running.  From there it performs
a binary search over the memory constraint until it finds the minimum
that allows for stable startup and health checks.

This new minimum of 345MB is probably too low for practical use, as some
additional memory would be required to run an actual flow, receive its
logs and events, and run periodic loop services.