Reduces the memory requirement of a standalone Prefect server again
In #16644, we were able to get the standalone Prefect server we run in
`prefect server start` from 400MB to 343MB.  By running `uvicorn`
directly in the parent process, we further reduce that to *266MB*.  I
measured this using the same process as before, using a `cgroup` with
no swap to constrain the memory until I found the minimum.

There are two main changes here:

1. Making `prefect server start` a synchronous command in both the
   background and the foreground cases.  This helps to simplify the
   handling of signals and removes the need for `anyio` or piping the
   streams of a child process manually.

2. Splitting the paths between background and foreground servers.  In
   the case of the background server, we run the subprocess as usual and
   then exit leaving it running.  In the case of the foreground server,
   we run it with `uvicorn.run` as the final call, letting `uvicorn`
   handle signals.  `uvicorn` already handles `SIGINT` and `SIGTERM` to
   cleanly shutdown, so we don't need to do anything additional here.

I'm removing the two tests that were testing what happens when the
signals are sent in escalating succession, as they aren't materially
different now.  In both cases, `uvicorn` will attempt to shutdown
cleanly.  I also updated the tests to reflect that `SIGINT` will lead to
a successful exit, while `SIGTERM` will lead to a -15 exit (pretty
standard).