Ensures that `QueueService` instances are unique within types
As @jakekaplan and I were debugging a different issue, we realized that
the implementation of `QueueServiceBase.instance` would return an
instance of the wrong type if there were two subclasses that both used
the same `args` (including the empty `args`).  Luckily, the subclasses
of `QueueService` we have so far seem to avoid this by overriding
`instance` or by using different `args`, but this leaves a trap for
folks implementing services in the future.