Adds `prefect-redis` to the Prefect image
In the interest of helping folks looking to horizontally scale out
Prefect, we want to include `prefect-redis` in the base image.  This is
unusual for a Prefect integration, but we feel that including the
Redis implementation of messaging and caching will enable folks to make
a simpler transition to running the server and services separately.
Since `redis` is already a dependency of Prefect, this will add very
minimal heft to the image (<1MB).