Trims Redis streams periodically while consuming
Redis streams are retained indefinitely, and they need to be purposely
trimmed to remove old messages.  For our message consumer case, we can
trim the stream as soon as all of its messages have been delivered and
acknowledged.  We'll do this periodically within the consumer so that
we don't need to run a separate loop service to do it.  This also means
that while a Prefect consumer is down for a period of time, its stream
won't be trimmed until it is started again to drain it.