fix: Implement batch deletion for events and event resources.
Even when using an indexed column, deleting event resource rows takes time - especially if this is the first time the table gets cleaned up and a billion old entries have to be deleted.

The events table does not necessarily need this logic, but for consistency is implemented the same way now.