Surfaces pull step errors as independent log messages
In trying to increase user's visibility into any problems with
dependency installations and code pulls, we want to make sure that the
output is clear and actionable in their logs.  Today, when we get a
step excution error, we are showing that as a full traceback, including
the underlying `__cause__`, which can be quite verbose for the
uninitiated.

Here we get the string representation of that exception message as a
separate `ERROR` log just before we re-raise it.  This comes out in the
flow run logs as a clean error, like:

```
`run_shell_script` failed with error code 1: Using Python 3.12.2 environment at: /home/chris/.pyenv/versions/orion-312
  × No solution found when resolving dependencies:
  ╰─▶ Because flibbidygibbles was not found in the package registry and you
      require flibbidygibbles>=9000, we can conclude that your requirements
      are unsatisfiable.
```