Brings back `EventPersister` in discovery of background services
In the prior work on this (PrefectHQ/prefect#16913) we needed to
maintain an explicit set of modules from which to discover services, and
I missed the modules that include the `EventPersister` and the
`EventLogger` (a debugging-only service).  This meant that under most
circumstances, because the `event_persister` module wasn't imported, we
wouldn't run it in either API or background processes.

Here I'm adding a test that confirms the exact subsets of services we
should run, just to avoid any quiet failures like this.