Include `pytz` in the `prefect-client` dependencies
If the `prefect-client` is loading a `Deployment` that includes any
`CronSchedule`s, it needs `pytz` for the vendor'd `croniter` to validate
those schedules.

This also adds a unit test that confirms that the `prefect-client`
dependencies that are defined match their corresponding dependency in
the main package (I only found that one that was out of sync).