Propagates the currently triggering event (if any) for composite triggers
Just like an event trigger, a composite trigger _may_ have a triggering
event that caused it to fire (like when all of its component triggers
are Reactive event triggers, a pretty common case).  Here, we propagate
that triggering event up through the firings so that folks can access
it from templates.  More importantly, this allows actions to make
inferences from the triggering event (like being able to infer a
deployment from it for the `PauseDeployment` action, for example).