Corrects and infinite loop in the `CancellationCleanup` service
This service wasn't keeping track of its progress, so as soon as you had
more than one `batch_size`'s worth of cancelled runs, it woudl be stuck
in an infinite loop.

Note: it would be nice to find a way to not revisit the same flow runs
over and over again with some additional bookkeeping, but this will at
least prevent the basic version of this problem.

Closes #18005