Fixes deadlock in task run recorder
These changes fix a deadlock that occurs when the task run recorder is inserting task runs.

The deadlock occurs when two tasks run events are persisted at the same time before the task run has been created in the DB. The `INSERT... ON CONFLICT DO UPDATE ...` statement causes both inserts to take speculative locks and wait for each other.

The fix is to use `INSERT... ON CONFLICT DO NOTHING ...` instead of `INSERT... ON CONFLICT DO UPDATE ...` and then check the result of the insert. If the insert did not result in a row being inserted, we update the task run instead.