Fix errors in ECS worker related to ad hoc submission
There are a couple of places in the ECS worker where we assume that a deployment ID is present. Flows submitted via the `@ecs` decorator don't have an assoicated deployment, which causes errors in caching, family name generation and log prefix generation. These changes adjust the implementation to use the flow ID as a fallback when there is no associated deployment.