fix: allow automations notify when triggers service is disabled
For maximum performance from the API servers, we might want to delegate
triggers/actions processing exclusively to background services.
In that case, we will disable the triggers service on the API servers.
But this will prevent automations change notifications to reach the
background services.

Allow this by removing this old check.

Signed-off-by: Fatih Acar <fatih@opsmill.com>