Fixes async locking error by generating unique holder for each transaction
These changes fix an error where running nested transactions across asyncio tasks caused an error when releasing a transactions lock. The default holder string contains the ID of the current task, so mixing tasks would result in mismatched holders on lock acquisition and release.

The fix is to generate a unique holder for each transaction, which is used to acquire and release the lock, and held for the duration of the transaction.

Closes https://github.com/PrefectHQ/prefect/issues/18600